useDeferred for keywords parsing
