+++ src/codex/hooks/useSearchFilters/index.ts
@@ -33,6 +33,7 @@ import { useExtraFilters } from './useExtraFilters'
 import { useFormattingFilters } from './useFormattingFilters'
 import { useCardStrike } from './useCardStrike'
 import { useTierFilters } from './useTierFilters'
+import { useKeywords } from './useKeywords'
 
 export interface UseCardSearchFilters {
   keywords: string
@@ -66,9 +67,12 @@ export interface UseTalentSearchFilters {
 export const useCardSearchFilters = (cardData: CardData[] | undefined): UseCardSearchFilters => {
   const cachedFilters = getCachedCardCodexSearchFilters()
   const { filterData, isFilterDataError, isFilterDataLoading } = useWeeklyChallengeFilterData()
-
-  const [keywords, setKeywordsUntracked] = useState(cachedFilters?.keywords || '')
-  const [parsedKeywords, setParsedKeywords] = useState<string[]>([])
+  const {
+    keywords,
+    setKeywords: setKeywordsUntracked,
+    parsedKeywords,
+    resetParsedKeywords,
+  } = useKeywords(cachedFilters?.keywords)
   const [matchingCards, setMatchingCards] = useState<CardData[]>([])
 
   const untrackedUseCardSetFilters = useCardSetFilters(cachedFilters?.cardSets)
@@ -151,7 +155,7 @@ export const useCardSearchFilters = (cardData: CardData[] | undefined): UseCardS
 
   const resetFilters = () => {
     trackedSetKeywords('')
-    setParsedKeywords([])
+    resetParsedKeywords()
     resetCardSetFilters()
     resetRarityFilters()
     resetBannerFilters()
@@ -222,12 +226,6 @@ export const useCardSearchFilters = (cardData: CardData[] | undefined): UseCardS
   // --------------------------------------------------
 
   useEffect(() => {
-    const parsed = parseKeywords(keywords)
-
-    if (!isArrayEqual(parsedKeywords, parsed)) {
-      setParsedKeywords(parsed)
-    }
-
     if (cardData) {
       const filteredCards = cardData
         .filter((card) =>
@@ -251,7 +249,7 @@ export const useCardSearchFilters = (cardData: CardData[] | undefined): UseCardS
 
           return true
         })
-        .filter((card) => isNameOrDescriptionIncluded(card, parsed))
+        .filter((card) => isNameOrDescriptionIncluded(card, parsedKeywords))
 
       if (!isArrayEqual(filteredCards, matchingCards, 'name')) {
         setMatchingCards(filteredCards)
@@ -262,9 +260,8 @@ export const useCardSearchFilters = (cardData: CardData[] | undefined): UseCardS
     isBannerIndexSelected,
     isCardSetIndexSelected,
     isRarityIndexSelected,
-    keywords,
-    matchingCards,
     parsedKeywords,
+    matchingCards,
     shouldIncludeMonsterCards,
     shouldIncludeNonCollectibleCards,
   ])
@@ -295,9 +292,12 @@ export const useTalentSearchFilters = (
   talentTree: TalentTree | undefined
 ): UseTalentSearchFilters => {
   const cachedFilters = getCachedTalentCodexSearchFilters()
-
-  const [keywords, setKeywordsUntracked] = useState(cachedFilters?.keywords || '')
-  const [parsedKeywords, setParsedKeywords] = useState<string[]>([])
+  const {
+    keywords,
+    setKeywords: setKeywordsUntracked,
+    parsedKeywords,
+    resetParsedKeywords,
+  } = useKeywords(cachedFilters?.keywords)
   const [matchingTalentTree, setMatchingTalentTree] = useState<TalentTree | undefined>(undefined)
 
   // TODO: Extras with Offers, talents from invasion events?
@@ -339,7 +339,7 @@ export const useTalentSearchFilters = (
 
   const resetFilters = () => {
     trackedSetKeywords('')
-    setParsedKeywords([])
+    resetParsedKeywords()
     resetCardSetFilters()
     resetTierFilters()
   }
@@ -400,17 +400,11 @@ export const useTalentSearchFilters = (
   }, [])
 
   useEffect(() => {
-    const parsed = parseKeywords(keywords)
-
-    if (!isArrayEqual(parsedKeywords, parsed)) {
-      setParsedKeywords(parsed)
-    }
-
     if (talentTree) {
       const filterFn = (node: TalentTreeTalentNode) =>
         isCardSetIndexSelected(node.expansion) &&
         isTierIndexSelected(node.tier) &&
-        isNameOrDescriptionIncluded(node, parsed)
+        isNameOrDescriptionIncluded(node, parsedKeywords)
 
       const filteredTalentNodes = talentTree.noReqNode.children
         .map((node) => filterTalentTreeNode(node, filterFn))
@@ -441,7 +435,6 @@ export const useTalentSearchFilters = (
     talentTree,
     isCardSetIndexSelected,
     isTierIndexSelected,
-    keywords,
     parsedKeywords,
     matchingTalentTree,
     filterTalentTreeNode,
@@ -459,12 +452,6 @@ export const useTalentSearchFilters = (
     resetFilters,
   }
 }
-
-const parseKeywords = (keywords: string): string[] =>
-  keywords
-    .split(/,\s+or\s+|,\s*|\s+or\s+/)
-    .map((keyword) => keyword.trim())
-    .filter(Boolean)
 
 const isNameOrDescriptionIncluded = (
   { name, description }: CardData | TalentTreeTalentNode,
